Revert duck-typing and re-introduce the authentication service interface
